# public/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel Swagger API Documentation'
  description: 'Backend API for the salon SaaS platform (Laravel + PostgreSQL) with JWT auth and tenant-scoped operations.'
  version: 1.0.0
servers:
  -
    url: https://admin.vynhq.com
    description: Production
  -
    url: http://127.0.0.1:8000
    description: Local (php artisan serve)
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Use `/api/auth/login` (or related auth endpoints) to obtain a JWT token. Include `X-Tenant` for tenant-scoped endpoints. Authentication is configured globally in the OpenAPI security scheme.'
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    default: []
    bearerAuth: []
paths:
...
  /api/admin/tenants:
    get:
      summary: ''
      operationId: getApiAdminTenants
      description: ''
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
    post:
      summary: ''
      operationId: postApiAdminTenants
      description: ''
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
  '/api/admin/tenants/{tenant_id}':
    get:
      summary: ''
      operationId: getApiAdminTenantsTenant_id
      description: ''
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
    put:
      summary: ''
      operationId: putApiAdminTenantsTenant_id
      description: ''
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
    patch:
      summary: ''
      operationId: patchApiAdminTenantsTenant_id
      description: ''
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
    delete:
      summary: ''
      operationId: deleteApiAdminTenantsTenant_id
      description: ''
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
    parameters:
      -
        in: path
        name: tenant_id
        description: 'The ID of the tenant.'
        example: 2
        required: true
        schema:
          type: integer
...
  /api/analytics/franchise:
    get:
      summary: "Return per-branch KPIs for the authenticated salon owner's tenant."
      operationId: returnPerBranchKPIsForTheAuthenticatedSalonOwnersTenant
      description: "Query params (all optional):\n  from  Y-m-d  defaults to first day of current month\n  to    Y-m-d  defaults to today"
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
...